Medusa/tests/functional/aws_protection/cvsa_manager/test_plan_updater/requirements.txt
beautifulsoup4==4.12.2
atlassian-python-api==3.36.0
pandas==2.0.0
